Add files via upload
One of the most basic abilities of human being involves the grouping of similar objects to produce a classification. The idea of sorting similar things into categories is clearly a primitive one since early civilizations.
Classification, in its widest sense, is needed for the development of language, which consists of words which help us to recognize and discuss the different types of events, objects and people we encounter. As well as being a basic human conceptual activity, classification is also fundamental to most branches of science.
Clustering is performed to get insights into the data whose volume makes it problematic for analysis by humans. Due to this, clustering algorithms have emerged as meta learning tools for performing exploratory data analysis. A cluster is defined as a set of objects which have a higher degree of similarity to each other compared to objects not in the same set. Multiple measures have been proposed related to quantifying similarity such as euclidean distance, density in data space etc. making clustering a multi-objective optimization problem.

The K-Means algorithm terminates when either the same cluster means show up in the next iteration or if the same clusters are formed in the next iteration. If neither of this happens then the K-Means algorithm halts the clustering process abruptly after a number of pre-specified iterations. Thus an abrupt termination cuts off the possibility that the means may have converged at a later stage providing a fair chance for accurate formation of clusters.
The K-Means clustering approach takes a very large time to converge as the number of data values increases mainly because it recomputes the means until the same means show up in the next iteration.